requirements.txt
Django==1.7.3
#psycopg2==2.5.1
git+https://github.com/schallis/psycopg2-windows.git@win32-py27#egg=psycopg2
requests==2.5.1
dateutils==0.6.6
django-jsonfield==0.9.13
django-azure-storage==0.0.1
retrying==1.3.3
django-debug-toolbar==1.2.2
